+++ apps/client/src/pages/ai-results/ui/ai-results-page.tsx
@@ -1,10 +1,10 @@
 import { useMemo } from 'react';
 
-import AllMemoPage, {
-  type AllMemoPageHelpers,
-} from '@pages/all-memo/ui/all-memo-page';
-
 import { MOCK_MEMOS } from '@widgets/memo-list/ui/mock-memos';
+import {
+  MemoListView,
+  type MemoListViewHelpers,
+} from '@widgets/memo-list-view';
 
 const AiResultsPage = () => {
   //TODO: 실제 API 연동 후 수정
@@ -12,13 +12,16 @@ const AiResultsPage = () => {
     return MOCK_MEMOS.filter((memo) => memo.aiResult === true);
   }, []);
 
-  const handleAiCreateClick = (memoId: string, helpers: AllMemoPageHelpers) => {
-    helpers.setInitialSelectedId(memoId);
+  const handleAiCreateClick = (
+    memoId: string,
+    helpers: MemoListViewHelpers,
+  ) => {
+    helpers.selectFunction(memoId);
     helpers.setIsAiMode(true);
   };
 
   return (
-    <AllMemoPage
+    <MemoListView
       title="AI 결과물"
       initialMemos={aiResultMemos}
       onAiCreateClick={handleAiCreateClick}
