+++ apps/client/src/widgets/memo-list-view/index.ts
@@ -1,0 +1,4 @@
+export type { MemoListViewHelpers } from './hooks/use-memo-list-view';
+export { useMemoListView } from './hooks/use-memo-list-view';
+export type { MemoListViewProps } from './ui/memo-list-view';
+export { default as MemoListView } from './ui/memo-list-view';
